Update map jungle and add textures
All textures for in game HUD is added.

Map jungle is updated to match with map. Due to some problems when
copying the terrain from map jungle to map, the resulting of lightmap is
not as expected. So we move to use map jungle instead. Please check all
functionalities in this map if they are working correctly.

The lobby VS texture is also updated.

Please check the HUD in game for better positioning.